Refactor: 검색결과 컴포넌트 pages로 이동
