+++ src/pages/searchResultPage/SearchResultPageStyle.jsx
